requirements.txt
pytest==7.2.2
requests==2.28.2
jsonschema==4.17.3
python-dotenv==1.0.0